Move packages from `pkg` to `internal`
