+++ internal/controller/applicationhealthcomment_controller.go
@@ -23,8 +23,8 @@ import (
 	"github.com/argoproj/gitops-engine/pkg/health"
 	argocdcommenterv1 "github.com/int128/argocd-commenter/api/v1"
 	"github.com/int128/argocd-commenter/internal/controller/predicates"
-	"github.com/int128/argocd-commenter/pkg/argocd"
-	"github.com/int128/argocd-commenter/pkg/notification"
+	"github.com/int128/argocd-commenter/internal/argocd"
+	"github.com/int128/argocd-commenter/internal/notification"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
